bybit-bot/requirements.txt
pybit==5.11.0
pandas==2.2.2
pandas-ta==0.3.14b0
python-dotenv==1.0.1
numpy==1.24.4
matplotlib==3.8.4
mplfinance==0.12.10b0
streamlit==1.35.0
plotly==5.21.0
fastapi==0.111.0
uvicorn==0.30.1
orjson